Refactored duplicate code in MemoryManagers using generics.
